Refactor JSObject code
